ファイルを完全分離した。
