+++ src/roguelike/buttle.py
@@ -1,0 +1,3 @@
+class Buttle:
+    "戦闘"
+    pass
