+++ src/roguelike/character.py
@@ -1,0 +1,12 @@
+class Character:
+    """登場人物"""
+
+    def __init__(self, icon, message):
+        self.icon = icon
+        self.message = message
+
+    def talk(self):
+        pass
+
+    def draw(self, draw_side):
+        pass
