+++ src/roguelike/enemy.py
@@ -1,0 +1,2 @@
+class Enemy:
+    pass
