+++ src/roguelike/fieldview.py
@@ -1,0 +1,13 @@
+class FieldView:
+    "フィールドビュー"
+    def __init__(self):
+        pass
+
+    def is_movable(self):
+        pass
+
+    def get_message(self):
+        pass
+
+    def draw(self):
+        pass
