+++ src/roguelike/item.py
@@ -1,0 +1,2 @@
+class Item:
+    pass
