+++ src/roguelike/player_character.py
@@ -1,0 +1,3 @@
+class PlayerCharacter:
+    """マップ上での操作キャラクター"""
+    pass
